阶段性更新：后端数据切换到整形、性能优化、bug修复

+++ InfrastSim/TimeDriven/Helper.cs
@@ -1,6 +1,3 @@
-using System.Text;
-using System.Text.Json;
-
 namespace InfrastSim.TimeDriven;
 internal static class Helper {
     public static bool VisibleToLookupSkill(this OperatorBase op) {
@@ -11,8 +8,7 @@ internal static class Helper {
             return !op.IsExausted;
         }
     }
-    public static OperatorBase? FindOp(this FacilityBase facility, string opName)
-    {
+    public static OperatorBase? FindOp(this FacilityBase facility, string opName) {
         return facility.Operators.Where(op => op.Name == opName).FirstOrDefault();
     }
     /// <summary>
@@ -36,31 +32,26 @@ internal static class Helper {
     /// <summary>
     /// 适用于 设施中是否有xxx组干员, 考虑红脸
     /// </summary>
-    public static bool HasGroupMember(this FacilityBase facility, string group)
-    {
+    public static bool HasGroupMember(this FacilityBase facility, string group) {
         return facility.GroupMembers(group).Any();
     }
     /// <summary>
     /// 适用于 设施中xxx组干员的数量, 考虑红脸
     /// </summary>
-    public static int GroupMemberCount(this FacilityBase facility, string group)
-    {
+    public static int GroupMemberCount(this FacilityBase facility, string group) {
         return facility.GroupMembers(group).Count();
     }
-    public static int GetRealGoldProductionLine(this Simulator simu)
-    {
+    public static int GetRealGoldProductionLine(this Simulator simu) {
         return simu.ModifiableFacilities
                    .Where(fac => fac is ManufacturingStation manufacturing
                        && manufacturing.Product == Product.Gold).Count();
     }
-    public static int GetGoldProductionLine(this Simulator simu)
-    {
-        return (int)simu.ExtraGoldProductionLine + simu.GetRealGoldProductionLine();
+    public static int GetGoldProductionLine(this Simulator simu) {
+        return simu.ExtraGoldProductionLine + simu.GetRealGoldProductionLine();
     }
 
-    public static int PowerStationsCount(this Simulator simu)
-    {
-        return (int)simu.ExtraPowerStation + simu.PowerStations.Count();
+    public static int PowerStationsCount(this Simulator simu) {
+        return simu.ExtraPowerStation + simu.PowerStations.Count();
     }
     public static bool IsProduceGold(this ManufacturingStation manufacturing) {
         return manufacturing.Product == Product.Gold;
